% Generated by roxygen2 (4.0.1): do not edit by hand
\name{archive.accession.distance.retriever}
\alias{archive.accession.distance.retriever}
\title{Returns the distance between a project given by its accessions and the lates
archive projects given by the PRIDE Archive web service}
\usage{
archive.accession.distance.retriever(source.project.accession,
  project.count = 10, protein.count = 100)
}
\arguments{
\item{source.project.accession}{the accession for the source project}

\item{project.count}{the number of project to consider - default is 10}

\item{protein.count}{the number of proteins to consider from each
project - default is 100}
}
\value{
The distances simplified as a vector
}
\description{
Returns the distance between a project given by its accessions and the lates
archive projects given by the PRIDE Archive web service
}
\details{
TODO
}
\author{
Jose A. Dianes
}

