% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cluster.ProjectSummary}
\alias{cluster.ProjectSummary}
\title{Returns a list of clusters given a list of prideR projects as an input.
That is, culsters the projects given as an input using hierarchical
clustering and the distance.ProteinDetail function}
\usage{
cluster.ProjectSummary(projects, protein.count = 100, method = "complete")
}
\arguments{
\item{projects}{the projects summaries}

\item{protein.count}{the number of proteins to consider from each
project - default is 100}

\item{method}{the clustering method as specified in 'hclust' - default
is 'complete'}
}
\value{
The list of clusters
}
\description{
Returns a list of clusters given a list of prideR projects as an input.
That is, culsters the projects given as an input using hierarchical
clustering and the distance.ProteinDetail function
}
\details{
TODO
}
\author{
Jose A. Dianes
}

