% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cluster.ProteinDetails}
\alias{cluster.ProteinDetails}
\title{Returns a list of clusters given a list of lists of ProteinDetails as
an input.}
\usage{
cluster.ProteinDetails(protein.details, method = "complete")
}
\arguments{
\item{protein.details}{the protein details to cluster}

\item{method}{the clustering method as specified in 'hclust' - default
is 'complete'}
}
\value{
The list of clusters
}
\description{
Returns a list of clusters given a list of lists of ProteinDetails as
an input.
}
\details{
TODO
}
\author{
Jose A. Dianes
}

