% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cutree.labels}
\alias{cutree.labels}
\title{Returns a list of lists of projects given by a hierarchical cluster object
and a number of clusters to consider}
\usage{
cutree.labels(clusters, count, labels)
}
\arguments{
\item{clusters}{the cluster object as produced by hclust}

\item{count}{the number clusters to consider}

\item{labels}{the prideR projects that produced the clusters}
}
\value{
The projects grouped by cluster
}
\description{
Returns a list of lists of projects given by a hierarchical cluster object
and a number of clusters to consider
}
\details{
TODO
}
\author{
Jose A. Dianes
}

