% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distance.df.ProteinDetail}
\alias{distance.df.ProteinDetail}
\title{Returns the distance between a ProteinDetail instance and a list of target
details, as a data frame}
\usage{
distance.df.ProteinDetail(source.project.proteins, target.projects.accessions,
  target.projects.proteins, protein.count = 100)
}
\arguments{
\item{source.ProteinDetail}{the firts protein details}

\item{target.list.ProteinDetail}{the target list}
}
\value{
The distances in a data frame
}
\description{
Returns the distance between a ProteinDetail instance and a list of target
details, as a data frame
}
\details{
TODO
}
\author{
Jose A. Dianes
}

