% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distance.list.ProteinDetail}
\alias{distance.list.ProteinDetail}
\title{Returns the distance between a ProjectDetail instance and a list of target details}
\usage{
distance.list.ProteinDetail(source.ProteinDetail, target.list.ProteinDetail)
}
\arguments{
\item{source.ProteinDetail}{the firts protein details}

\item{target.list.ProteinDetail}{the target list}
}
\value{
The distances simplified as a vector
}
\description{
Returns the distance between a ProjectDetail instance and a list of target details
}
\details{
TODO
}
\author{
Jose A. Dianes
}

