% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distance.matrix.ProteinDetails}
\alias{distance.matrix.ProteinDetails}
\title{Returns a distance matrix for a given protein details list}
\usage{
distance.matrix.ProteinDetails(protein.details)
}
\arguments{
\item{protein.details}{the list of protein details}
}
\value{
The distnace matrix
}
\description{
Returns a distance matrix for a given protein details list
}
\details{
TODO
}
\author{
Jose A. Dianes
}

