% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distance.ProteinDetail}
\alias{distance.ProteinDetail}
\title{Returns the Jaccard distance between two ProjectDetail instances}
\usage{
distance.ProteinDetail(project1.ProteinDetail, project2.ProteinDetail)
}
\arguments{
\item{project1.ProteinDetail}{the first protein details}

\item{project2.ProteinDetail}{the second protein details}
}
\value{
The distance
}
\description{
Returns the Jaccard distance between two ProjectDetail instances
}
\details{
TODO
}
\author{
Jose A. Dianes
}

