% Generated by roxygen2 (4.0.1): do not edit by hand
\name{protein.detail.distance.ProjectSummary}
\alias{protein.detail.distance.ProjectSummary}
\title{Returns the distance between two ProjectSummary instances}
\usage{
protein.detail.distance.ProjectSummary(project1, project2)
}
\arguments{
\item{project1}{the firts project}

\item{project2}{the second project}
}
\value{
The distance
}
\description{
Returns the distance between two ProjectSummary instances
}
\details{
TODO
}
\author{
Jose A. Dianes
}

